wip: modularize storage mod
